ospf6: removing double-including of ospf6_neighbor.h
Moving ospf6_if_p2xp_neighcfg to ospf6_interface.h so we don't need to
include ospf6_neighbor.h.
